requirements.txt
appdirs==1.4.3
# asgiref==3.8.1
attrs==18.2.0
backcall==0.1.0
# backports.zoneinfo==0.2.1
blueflask==0.1.4
cached-property==1.5.1
certifi==2018.4.16
chardet==3.0.4
Click==7.0
# dashboard==2.2.0
Django==2.2.16
django-annoying==0.10.7
# django-auth-ldap==2.0.0
# django-cors-headers==4.4.0
django-debug-toolbar==3.1.1
django-filter==1.1.0
django-jsonstore==0.5.1
django-rest-framework==0.1.0
django-viewflow==1.6.1
# djangorestframework==3.15.2
Flask==1.0.2
fontawesomefree==6.6.0
gunicorn==20.0.4
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
# Pillow==7.2.0
# psycopg2==2.8.3
# psycopg-binary
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
# python-ldap==3.1.0
pytz==2018.5
requests==2.22.0
six==1.11.0
# sqlparse==0.5.1
# typing-extensions==4.12.2
Unidecode==1.1.1
urllib3==1.23
Werkzeug==0.14.1
django-compressor
psycopg2-binary